+++ apps/fluent-tester/src/TestComponents/CornerRadius/consts.ts
@@ -1,0 +1,2 @@
+export const HOMEPAGE_CORNERRADIUS_BUTTON = 'Homepage_CornerRadius_Button';
+export const HOMEPAGE_CORNERRADIUS_TESTPAGE = 'CornerRadius_TestPage';