+++ packages/experimental/ExperimentalIcon/src/FontIcon/FontIcon.types.ts
@@ -1,0 +1,12 @@
+import { ColorValue, TextStyle, StyleProp } from 'react-native';
+
+export const fontIconName = 'FontIcon';
+
+export interface FontIconProps {
+  codepoint: number;
+  color?: ColorValue;
+  fontFamily?: string;
+  fontSize?: number;
+  fontSrcFile?: string;
+  style?: StyleProp<TextStyle>;
+}